Cleanup examples
 - Repair lib/oml4r/oml4r-wlanconfig.rb
 - Also adjust oml4r-simple-example to use :int32

+++ ruby/oml4r/lib/oml4r/oml4r-simple-example.rb
@@ -34,7 +34,7 @@ class SinMP < OML4R::MPBase
   #channel :default
 
   param :label
-  param :angle, :type => :long
+  param :angle, :type => :int32
   param :value, :type => :double
 end
